[SIEM] Better URL state tests (#37745)
Better SIEM URL state tests
