+++ x-pack/plugins/siem/public/store/inputs/model.ts
@@ -4,6 +4,8 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 
+import { CONSTANTS } from '../../components/url_state/constants';
+
 export interface AbsoluteTimeRange {
   kind: 'absolute';
   fromStr: undefined;
@@ -53,7 +55,11 @@ export interface InputsModel {
   global: InputsRange;
   timeline: InputsRange;
 }
+export interface UrlInputsModelInputs {
+  linkTo: InputsModelId[];
+  [CONSTANTS.timerange]: TimeRange;
+}
 export interface UrlInputsModel {
-  global: TimeRange & LinkTo;
-  timeline: TimeRange & LinkTo;
+  global: UrlInputsModelInputs;
+  timeline: UrlInputsModelInputs;
 }
